+++ package.json
@@ -65,7 +65,7 @@
     "backstopjs": "^5.3.0",
     "c8": "^7.11.0",
     "chalk": "^4.1.0",
-    "chromedriver": "^113.0.0",
+    "chromedriver": "^115.0.0",
     "concat": "^1.0.3",
     "cross-env": "^7.0.3",
     "cz-conventional-changelog": "^3.3.0",
@@ -92,7 +92,6 @@
     "postcss-calc": "^8.0.0",
     "postcss-import": "^14.0.2",
     "postcss-inline-svg": "^5.0.0",
-    "postcss-jit-props": "^1.0.13",
     "postcss-load-config": "^3.0.1",
     "postcss-modules": "^4.0.0",
     "postcss-pxtorem": "^6.0.0",
