+++ src/components/rating/components/CdrStar00.vue
@@ -6,7 +6,7 @@
     xmlns="http://www.w3.org/2000/svg"
   >
     <path
-    :class="style['cdr-rating__star--empty']"
+      :class="style['cdr-rating__star--empty']"
       d="M12.1702 17.7643L11.9688 17.6359L11.7673 17.7643L6.37635 21.1984L7.81044
         14.6842L7.85695 14.473L7.69738 14.3269L2.8485 9.88945L9.21759 9.31204L9.44623
         9.29131L9.53174 9.07825L11.9688 3.00672L14.4058 9.07825L14.4913 9.29131L14.7199
