+++ src/components/rating/components/CdrStar100.vue
