+++ src/components/rating/components/CdrStar25.vue
