+++ src/components/rating/components/CdrStar50.vue
