+++ src/components/rating/components/CdrStar75.vue
